+++ R/check_no_rs_snp.R
@@ -18,7 +18,7 @@
  check_no_rs_snp <- function(sumstats_dt, path, ref_genome, snp_ids_are_rs_ids, indels,
                               imputation_ind, log_folder_ind, check_save_out,
                               tabix_index, nThread, log_files, dbSNP,
-                              dbSNP_tarball = NUL) {
+                              dbSNP_tarball) {
     SNP <- CHR <- CHR1 <- BP1 <- i.RefSNP_id <- IMPUTATION_SNP <-
         SNP_old_temp <- SNP_INFO <- A1 <- A2 <- .I <- NULL
     # if snp ids aren't rs ids rename the column to ID's 
@@ -93,6 +93,7 @@
                         nThread = nThread, 
                         log_files = log_files, 
                         dbSNP = dbSNP,
+                        dbSNP_tarball = dbSNP_tarball,
                         verbose = FALSE
                     )
                 log_files <- corrected_snp$log_files
@@ -171,10 +172,10 @@
             )
             message(msg)
              SNP_LOC_DATA <- load_snp_loc_data(
-                 ref_genome    = ref_genome,
-                 dbSNP   = dbSNP,
+                 ref_genome = ref_genome,
+                 dbSNP = dbSNP,
                  dbSNP_tarball = dbSNP_tarball,
-                 msg           = NULL
+                 msg = NULL
                )
             # split out chr:bp - check if chr or bp first by longer of two
             splits <- strsplit(miss_rs_chr_bp[1, SNP],
@@ -311,7 +312,8 @@
                                                         nchar(A2)>1),])
                         if(num_indels>0){
                             msg <- paste0("Found ",
-                                          formatC(nrow(num_indels),big.mark = ","),
+                                          formatC(nrow(num_indels),
+                                                  big.mark = ","),
                                           " Indels. These won't",
                                           " be checked against the referenc",
                                           "e genome as it does not contain ",
@@ -336,6 +338,7 @@
                             nThread = nThread, 
                             log_files = log_files,
                             dbSNP=dbSNP,
+                            dbSNP_tarball = dbSNP_tarball,
                             verbose = FALSE
                         )
                     log_files <- corrected_snp$log_files