+++ R/check_no_snp.R
@@ -13,12 +13,10 @@
 #' @importFrom data.table copy
 #' @importFrom BSgenome snpsByOverlaps
 #' @importFrom GenomicRanges makeGRangesFromDataFrame
-check_no_snp <- function(sumstats_dt, path, ref_genome, snp_ids_are_rs_ids, indels,
-                                                   imputation_ind, log_folder_ind, check_save_out,
-                                                   tabix_index, nThread, log_files, dbSNP,
-                                                   dbSNP_tarball = NULL,
-                                                   msg = NULL,
-                                                   verbose = TRUE) {
+check_no_snp <- function(sumstats_dt, path, ref_genome, snp_ids_are_rs_ids, 
+                         indels, imputation_ind, log_folder_ind, check_save_out,
+                         tabix_index, nThread, log_files, dbSNP,
+                         dbSNP_tarball = NULL,msg = NULL,verbose = TRUE) {
     SNP <- CHR <- i.RefSNP_id <- IMPUTATION_SNP <- BP <- A1 <- A2 <- NULL
     # If CHR and BP are present BUT not SNP then need 
     # to find the relevant SNP ids
@@ -30,10 +28,10 @@ check_no_snp <- function(sumstats_dt, path, ref_genome, snp_ids_are_rs_ids, inde
             msg <- NULL
         }
         SNP_LOC_DATA <- load_snp_loc_data(
-          ref_genome    = ref_genome,
-          dbSNP   = dbSNP,
+          ref_genome = ref_genome,
+          dbSNP = dbSNP,
           dbSNP_tarball = dbSNP_tarball,
-          msg           = NULL
+          msg = NULL
         )
         # if chromosome col has chr prefix remove it
         sumstats_dt[, CHR := gsub("chr", "", CHR)]