+++ R/get_genome_builds.R
@@ -17,6 +17,11 @@
 #' from its respective file path.
 #' Only works if \code{sumstats_list} is a list of paths.
 #' @param dbSNP version of dbSNP to be used (144 or 155). Default is 155.
+#' @param dbSNP_tarball Pass local versions of dbSNP in tarball format. Default 
+#' of NULL uses the dbSNP version passed in `dbSNP` parmeter. `dbSNP_tarball` 
+#' was enabled to help with dbSNP versions >=156, after the decision to no 
+#' longer provide dbSNP releases as bioconductor packages. dbSNP 156 tarball is 
+#' available here: http://149.165.171.124/SNPlocs/.
 #' @param nThread Number of threads to use for parallel processes.
 #' @param chr_filt Internal for testing - filter reference genomes and sumstats
 #' to specific chromosomes for testing. Pass a list of chroms in format: 
@@ -55,6 +60,7 @@ get_genome_builds <- function(sumstats_list,
                               sampled_snps = 10000,
                               names_from_paths = FALSE,
                               dbSNP=155,
+                              dbSNP_tarball=NULL,
                               nThread = 1,
                               chr_filt = NULL) {
     start <- Sys.time()
@@ -92,6 +98,7 @@ get_genome_builds <- function(sumstats_list,
         sumstats = sumstats_list[[1]],
         sampled_snps = sampled_snps,
         dbSNP = dbSNP, 
+        dbSNP_tarball = dbSNP_tarball,
         header_only = header_only,
         nThread = nThread,
         chr_filt = chr_filt
@@ -103,12 +110,14 @@ get_genome_builds <- function(sumstats_list,
                                    function(x,
                                             .sampled_snps = sampled_snps,
                                             .dbSNP=dbSNP,
+                                            .dbSNP_tarball = dbSNP_tarball,
                                             .header_only = header_only) {
                                      message_parallel(x)
                                      get_genome_build(
                                        sumstats = sumstats_list[[x]],
                                        sampled_snps = .sampled_snps,
                                        dbSNP = .dbSNP, 
+                                       dbSNP_tarball = .dbSNP_tarball,
                                        header_only = .header_only,
                                        nThread = 1,
                                        chr_filt = chr_filt